+++ DUBSON_Googs_Delivery_Program/TransportInterface.cs
@@ -6,7 +6,7 @@ namespace DUBSON_Goods_Delivery_Program
 {
     public interface TransportInterface
     {
-        public long Speed { get; }
+        public double Speed { get; }
 
         public bool isCurrentlyBusy { get; }
